requirements.txt
# VidInsight - B站视频智能笔记助手 依赖列表
# 时间: 2026-01-14
# @author: zhoujunyu

# Web UI 框架
streamlit>=1.28.0

# 视频/音频下载
yt-dlp>=2023.11.16

# HTTP 请求（用于调用远程 Whisper API）
requests>=2.31.0

# LLM API 调用
openai>=1.3.0

# 环境变量管理
python-dotenv>=1.0.0

# 思维导图渲染
streamlit-markmap>=0.1.0

# 其他工具
ffmpeg-python>=0.2.0
streamlit-cookies-manager>=0.0.1
